requirements.txt
streamlit
tensorflow
numpy
pillow
plotly
pandas